Factored out some classes from VerificationServer
